refactor: #62-EventList 뷰컨트롤러에 통합
