+++ EventLogger/Scenes/EventList/EventListReactor.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftData
-
 import Dependencies
 import ReactorKit
 import RxFlow
